extract types to separate files and tweak copy on warning pane
